+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- filtering messages by mailer, action and send_to
+
 ## [0.3.0] - 2022-10-04
 ### Added
 - UI for rendering dashboard with email tracking statistics